Transition to using minted package for listings
This commit Resolves #7 by replacing use of the `listings` package in
favor of the `minted` package. The new package is much more reliable for
linting code, however it requires that the user has Python 2.6+
installed on their system. This is not a problem for Overleaf, but might
pose compatibility issues for users building the document locally.

Summary of Changes:
- Replaced `listings` package with `minted` package
- Added .gitignore entry for minted cache folder
- Created a simplified listings example and moved it into the `examples`
  folder
- Added `RunCase.m` to serve as a better code example for the example
  listing
- Moved `listings` configuration from `Appendices` into
  `tudelft-light/compat` in case the listings package is required in the
  future for compatibility reasons
- Removed old Appendices folder