Fix typo in CHANGELOG and CICD errors (#20)
This commit fixes a typo int he CHANGELOG as well as fixes errors in the
CICD pipeline caused by CHANGELOG linting. A new GitHub action is
adopted in the workflow. Finally, the build workflow is restricted to
only running on pushes to the master branch and during pull requests
to reduce service utilization.