Add title fontsize option and remove hyphenation
This commit addresses Issue #23 by pulling new changes from the
tudelft-light submodule. These changes give the user the option to
specify a title fontsize for the cover page as well as disabling
hyphenation using the hyphenat package.

The README is updated to include the new title fontsize option and
a new minor release is added to the CHANGELOG

Summary of Changes:
- Updated tudelft-light submodule
- Added title fontsize argument to the README
- Updated CHANGELOG for minor release v2.2.0