Fix namespace conflict caused by report.cls
This commit Closes #34 by prepending tud- to all template
packages and class files. As a result, the namespace conflict which
led to failing compilations on TeXLive 2020 is fixed.

In addition, this commit Closes #31 by adding an  optional value
argument to \nomunit. The solution is to add an optional argument to the
\nomunit macro to pass this value onto \SI as opposed to the \si
environment. This fix is included in Commit [9628f4e4].

[9628f4e4]: https://github.com/skilkis/tudelft-light-template/commit/9628f4e4